Assignment #4 part #10 all done except for one small autorotation problem

+++ Matchismo/CardMatchingGame.m
@@ -56,7 +56,7 @@ static const int DEFAULT_MATCH_NUM = 2;
 // setting the score back to 0
 - (void)newGame:(Deck *)deck {
     _deck = deck;
-    int count = [self.cards count];
+    NSUInteger count = [self.cards count];
     NSMutableArray *discardItems = [NSMutableArray array];
     
     // remove cards in cards array
